Added /logfile= option to uninstaller.
Usage: Uninst.exe /logfile="c:\path\to\logfile".

Changed function DumpLog to JAWSDumpLog, changed its body to a macro, added function un.JAWSDumpLog.  Added macro JAWSDumpUninstLog which dumps log in un.OnUnInstSuccess and un.OnUninstFailed.  Added un.OnInstFailed which currently only prints a message and dumps the log if requested.

When the installer uninstalls an existing install, it DetailPrints the uninstaller log to the log.  However, since it is a silent uninstall, there is no window to get messages from so no log is dumped.

Now deletes the temp copy of the uninstaller.