We now always log the install info file install.ini.
This was not being logged when installing over an existing install and therefore for ever  not removed.

Made references to SecInstDirFiles conditioned on JAWSInstallFullItems being defined.  This removes a warning produced if the user does not define this macro.

This must not be used with uninstlog with commits after db91f5c.  Commits after this point implement uninstaller logging (v0.1.1).