Added definitions of JAWSLang_XXX defines.
If none of them are defined, all of them are defined.

Added defines in vwapp.nsi.

Added (commented out) test LangString to test just English support and Installer-specific section.

Modified instructions for adding a language in README.

Tested with all defines commented out and with English defined and 1 English LangString.