Smarter check for existing scripts.
Previously, when the installer checked for existing scripts in the script folders, it checked for anything in appname.* except a JCF file.  Now it checks for the script extensions explicitly, so it won't complain about something like app.jsd~.

Changed function CheckScriptExists to JawsCheckScriptExists.
Fixed comment about return from JawsCheckScriptExists.
Changed so that JawsCheckScriptExists does not write message, moved message to DisplayJawsListLeave.  This is so that it can be used in PageInstConfirmPre.

Note: Message InstConfirmHaveFiles says that files matching ${ScriptApp}.* may be overwritten.  This should be changed to something like
"$0The following JAWS versions contain files for this application: $1$\r$\nThese files will be overwritten during installation.$\r$\n"

It is not being changed now for the sake of translations.  It is not wrong, but it could be better.

Updated package version in README, no other changes to README.

Changed package rev to 2.6.