Made definition of NSD_GetStyle and macro conditional on not defined.
NSIS V3.03 provides a NSD_GetStyle macro and mine conflicts with it.
Also changed my version to be compatible with the provided one.
Changed invocation to match, but not tested.
This macro does not appear to be used in any uncommented code.

Executes build b i successfully with NSIS 3.03.

Package rev unchanged as to 2.8.